refactored StateMachine code into a new package + started refactoring autonomous poses to only have 1 set of poses which we flip
