+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/StateMachine/StateMachine.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.tools;
+package org.firstinspires.ftc.teamcode.tools.StateMachine;
 
 import java.util.ArrayList;
 import java.util.function.BooleanSupplier;
