+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tools/StateMachine/Timer.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.tools;
+package org.firstinspires.ftc.teamcode.tools.StateMachine;
 
 public class Timer {
     Timer(String name) {
